// src/frontend/webapp/modilist/src/services/swagger/api/models/upsert-style-preferences.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ModilistAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface UpsertStylePreferences
 */
export interface UpsertStylePreferences {
    /**
     * 
     * @type {string}
     * @memberof UpsertStylePreferences
     */
    choiceReasons?: string;
    /**
     * 
     * @type {number}
     * @memberof UpsertStylePreferences
     */
    lovesShopping?: number;
    /**
     * 
     * @type {number}
     * @memberof UpsertStylePreferences
     */
    openToSuggestions?: number;
    /**
     * 
     * @type {boolean}
     * @memberof UpsertStylePreferences
     */
    prefersHijabClothing?: boolean;
    /**
     * 
     * @type {string}
     * @memberof UpsertStylePreferences
     */
    bodyPartsToHighlight?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpsertStylePreferences
     */
    bodyPartsToHide?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpsertStylePreferences
     */
    excludedUpperCategories?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpsertStylePreferences
     */
    excludedLowerCategories?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpsertStylePreferences
     */
    excludedOuterCategories?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpsertStylePreferences
     */
    excludedUnderwearCategories?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpsertStylePreferences
     */
    excludedAccessoryCategories?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpsertStylePreferences
     */
    excludedFootwearCategories?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpsertStylePreferences
     */
    excludedBagCategories?: string | null;
}